modules/examples/mlops-stepfunctions/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
annotated-types==0.7.0
    # via pydantic
attrs==23.2.0
    # via
    #   cattrs
    #   jsii
    #   jsonschema
    #   referencing
    #   sagemaker
aws-cdk-asset-awscli-v1==2.2.202
    # via aws-cdk-lib
aws-cdk-asset-kubectl-v20==2.1.2
    # via aws-cdk-lib
aws-cdk-asset-node-proxy-agent-v6==2.0.3
    # via aws-cdk-lib
aws-cdk-aws-lambda-python-alpha==2.153.0a0
    # via -r requirements.in
aws-cdk-lib==2.153.0
    # via
    #   -r requirements.in
    #   aws-cdk-aws-lambda-python-alpha
    #   cdk-nag
boto3==1.35.1
    # via
    #   -r requirements.in
    #   sagemaker
botocore==1.35.1
    # via
    #   boto3
    #   s3transfer
cattrs==22.1.0
    # via jsii
cdk-nag==2.28.185
    # via -r requirements.in
certifi==2024.7.4
    # via requests
charset-normalizer==3.3.2
    # via requests
cloudpickle==2.2.1
    # via sagemaker
constructs==10.1.37
    # via
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-lib
    #   cdk-nag
dill==0.3.8
    # via
    #   multiprocess
    #   pathos
docker==7.1.0
    # via sagemaker
exceptiongroup==1.2.2
    # via cattrs
google-pasta==0.2.0
    # via sagemaker
idna==3.7
    # via requests
importlib-metadata==6.11.0
    # via sagemaker
importlib-resources==6.4.0
    # via jsii
jmespath==1.0.0
    # via
    #   boto3
    #   botocore
joblib==1.4.2
    # via scikit-learn
joblib-stubs==1.4.2.3.20240619
    # via -r requirements.in
jsii==1.102.0
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-lib
    #   cdk-nag
    #   constructs
jsonschema==4.23.0
    # via sagemaker
jsonschema-specifications==2023.12.1
    # via jsonschema
multiprocess==0.70.16
    # via pathos
numpy==1.26.4
    # via
    #   pandas
    #   pandas-stubs
    #   sagemaker
    #   scikit-learn
    #   scipy
packaging==24.1
    # via sagemaker
pandas==2.2.2
    # via
    #   -r requirements.in
    #   sagemaker
pandas-stubs==2.2.2.240807
    # via -r requirements.in
pathos==0.3.2
    # via sagemaker
platformdirs==4.2.2
    # via sagemaker
pox==0.3.4
    # via pathos
ppft==1.7.6.8
    # via pathos
protobuf==4.25.4
    # via sagemaker
psutil==6.0.0
    # via sagemaker
publication==0.0.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-lib
    #   cdk-nag
    #   constructs
    #   jsii
pydantic==2.7.4
    # via
    #   -r requirements.in
    #   pydantic-settings
pydantic-core==2.18.4
    # via pydantic
pydantic-settings==2.3.3
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   botocore
    #   jsii
    #   pandas
python-dotenv==1.0.1
    # via pydantic-settings
pytz==2024.1
    # via pandas
pyyaml==6.0.2
    # via sagemaker
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   docker
    #   sagemaker
rpds-py==0.20.0
    # via
    #   jsonschema
    #   referencing
s3transfer==0.10.1
    # via boto3
sagemaker==2.229.0
    # via -r requirements.in
schema==0.7.7
    # via sagemaker
scikit-learn==1.5.1
    # via -r requirements.in
scipy==1.13.1
    # via scikit-learn
six==1.16.0
    # via
    #   google-pasta
    #   python-dateutil
smdebug-rulesconfig==1.0.1
    # via sagemaker
tblib==3.0.0
    # via sagemaker
threadpoolctl==3.5.0
    # via scikit-learn
tqdm==4.66.5
    # via sagemaker
typeguard==2.13.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-lib
    #   cdk-nag
    #   jsii
types-pytz==2024.1.0.20240417
    # via pandas-stubs
types-pyyaml==6.0.12.20240808
    # via -r requirements.in
typing-extensions==4.7.1
    # via
    #   joblib-stubs
    #   jsii
    #   pydantic
    #   pydantic-core
tzdata==2024.1
    # via pandas
urllib3==1.26.19
    # via
    #   botocore
    #   docker
    #   requests
    #   sagemaker
zipp==3.20.0
    # via
    #   importlib-metadata
    #   importlib-resources
